Addressed comments.
1. Groupped kernel functions into two
2. Moved `pack_and_push` and `pack_and_push_block` to `program_utils.hpp`
3. Removed unnecessary comment and improved already exisiting one
   regarding writting `dgamma_components`
