+++ tt-train/.vscode/settings.json
@@ -1,0 +1,5 @@
+{
+    "editor.formatOnSave": true,
+    "files.autoSave": "afterDelay",
+    "C_Cpp.clang_format_style": ".clang-format"
+}
